Make str_repeat call stricter
This was causing an issue with PhpSpec

+++ src/Exporter.php
@@ -240,7 +240,7 @@ class Exporter
             "'";
         }
 
-        $whitespace = \str_repeat(' ', 4 * $indentation);
+        $whitespace = \str_repeat(' ', (int)(4 * $indentation));
 
         if (!$processed) {
             $processed = new Context;
